requirements.txt
Django==4.1.3
Pillow
djangorestframework==3.14.0
markdown
opencv-python-headless
django-sslserver